Adapt the human input portion of swerve-example/Robot.java into a subsystem.
Along the way, I've fixed (or at least I think I've fixed) some
long-standing flaws other our team's input subsystems have had:

- We don't just check for joysticks during construction -- we do it
  periodically.
- We check for disconnections, too (and for *being* disconnected.)

All of this means that exceptions from the controllers should be a thing
of the past: If you forget to plug in the XBox controller, plug it in
and use it -- no need to restart the robot.

Still missing NetworkTables for keyboard support.  Gotta fix
NetworkTablesInput for that.