Removed unused code.
These swerve module backends are no longer part of the system architecture.

The original intent was to implement multiple types of swerve module using two
types of encoders -- the 2020bot scheme with relative encoders and the new
CANCoder scheme.  In practice, we are only doing the second one, that code was
baked into the DriveSubsystem, and the bot going to competition next week will
be using a differential drive anyway.