Reduce lift speed; make separate left/right configs for the arm.
The arm lift axle twisted today due to the two brushless NEO motors spinning
full speed in opposing directions.  This is an expensive repair, and it would
not have been necessary if we had done things right to begin with.  I share
much of the blame for this; we discussed the possibility of this happening but
effected no code changes.

There are now four config objects instead of two.  The only difference between
the left and right configurations is the follow() command.  Left always leads
right; right always spins at the same magnitude as left, but in the opposite
direction from left.  Since they are mirrored on each side of the arm, this
should make the motors work together.

Emphasis on 'should.'